UCPT: conversion steps widget created (except loading saved data while edit question)
